Change heating_status to uint8_t
Saves 66 byte of flash and 1 byte of SRAM

+++ Firmware/Marlin.h
@@ -322,7 +322,7 @@ extern bool loading_flag;
 extern uint8_t usb_printing_counter;
 extern unsigned long total_filament_used;
 void save_statistics(unsigned long _total_filament_used, unsigned long _total_print_time);
-extern unsigned int heating_status;
+extern uint8_t heating_status;
 extern unsigned int status_number;
 extern unsigned int heating_status_counter;
 extern char snmm_filaments_used;
